# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Distill API",
    "version": "0.1.0"
  },
  "paths": {
    "/requests": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Submit a feature request",
        "description": "Saves the request as `pending` and returns at once; the intake workflow runs later.",
        "operationId": "create_request_requests_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestOut"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      },
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "A requester's own requests, newest first, with their status and need",
        "operationId": "my_requests_requests_get",
        "parameters": [
          {
            "name": "requester_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Requester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyRequest"
                  },
                  "title": "Response My Requests Requests Get"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/needs": {
      "get": {
        "tags": [
          "needs"
        ],
        "summary": "Find needs",
        "operationId": "list_needs_needs_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text in the need or its member requests",
              "title": "Q"
            },
            "description": "Text in the need or its member requests"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NeedStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "product_area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Area"
            }
          },
          {
            "name": "segment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Segment"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Needs with a request or confirmed support from it",
              "title": "Segment"
            },
            "description": "Needs with a request or confirmed support from it"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "priority",
                "support",
                "recent"
              ],
              "type": "string",
              "default": "priority",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeedPage"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/needs/{need_id}": {
      "get": {
        "tags": [
          "needs"
        ],
        "summary": "Get Need",
        "operationId": "get_need_needs__need_id__get",
        "parameters": [
          {
            "name": "need_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Need Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeedDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/portal/needs": {
      "get": {
        "tags": [
          "portal"
        ],
        "summary": "Find needs (requester view: no scores, revenue or accounts)",
        "operationId": "list_needs_portal_needs_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NeedStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "product_area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Area"
            }
          },
          {
            "name": "segment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Segment"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Segment"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "priority",
                "support",
                "recent"
              ],
              "type": "string",
              "default": "priority",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalNeedPage"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/portal/needs/{need_id}": {
      "get": {
        "tags": [
          "portal"
        ],
        "summary": "A need as a requester sees it, with the updates written to them",
        "operationId": "need_portal_needs__need_id__get",
        "parameters": [
          {
            "name": "need_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Need Id"
            }
          },
          {
            "name": "requester_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 9223372036854775807,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Whose updates to include",
              "title": "Requester Id"
            },
            "description": "Whose updates to include"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalNeedDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "MyRequest": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "needs_review_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needs Review Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "need": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NeedRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "needs_review_reason",
          "created_at",
          "processed_at",
          "need"
        ],
        "title": "MyRequest"
      },
      "NeedDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "product_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Area"
          },
          "status": {
            "$ref": "#/components/schemas/NeedStatus"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score",
            "description": "0-100, computed when read from current data and config"
          },
          "breakdown": {
            "$ref": "#/components/schemas/PriorityBreakdown"
          },
          "support_count": {
            "type": "integer",
            "title": "Support Count",
            "description": "Distinct requesters with a member request or a confirmed support. Staff filing for several customers count once; account_count shows the customers."
          },
          "claimed_support_count": {
            "type": "integer",
            "title": "Claimed Support Count",
            "description": "Supports still waiting for the intake workflow to confirm them"
          },
          "request_count": {
            "type": "integer",
            "title": "Request Count"
          },
          "account_count": {
            "type": "integer",
            "title": "Account Count"
          },
          "last_request_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Request At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "job_to_be_done": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job To Be Done"
          },
          "merged_into_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merged Into Id"
          },
          "requests": {
            "items": {
              "$ref": "#/components/schemas/NeedRequestOut"
            },
            "type": "array",
            "title": "Requests"
          },
          "supports": {
            "items": {
              "$ref": "#/components/schemas/SupportOut"
            },
            "type": "array",
            "title": "Supports"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountOut"
            },
            "type": "array",
            "title": "Accounts"
          },
          "origin": {
            "$ref": "#/components/schemas/NeedOrigin"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "updates": {
            "items": {
              "$ref": "#/components/schemas/UpdateOut"
            },
            "type": "array",
            "title": "Updates",
            "description": "Approved updates sent to supporters (drafts are never shown)"
          },
          "audit_trail": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "type": "array",
            "title": "Audit Trail",
            "description": "Links, unlinks, status changes, decisions and AI runs, newest first"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "problem",
          "persona",
          "product_area",
          "status",
          "priority_score",
          "breakdown",
          "support_count",
          "claimed_support_count",
          "request_count",
          "account_count",
          "last_request_at",
          "created_at",
          "job_to_be_done",
          "merged_into_id",
          "requests",
          "supports",
          "accounts",
          "origin",
          "evidence",
          "updates",
          "audit_trail"
        ],
        "title": "NeedDetail"
      },
...
      "NeedPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NeedSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "NeedPage"
      },
      "NeedRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "product_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Area"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "problem",
          "persona",
          "product_area"
        ],
        "title": "NeedRef"
      },
...
      "NeedStatus": {
        "type": "string",
        "enum": [
          "open",
          "planned",
          "in_progress",
          "shipped",
          "declined",
          "merged"
        ],
        "title": "NeedStatus"
      },
...
      "PortalNeedDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "product_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Area"
          },
          "status": {
            "$ref": "#/components/schemas/NeedStatus"
          },
          "support_count": {
            "type": "integer",
            "title": "Support Count"
          },
          "account_count": {
            "type": "integer",
            "title": "Account Count"
          },
          "request_count": {
            "type": "integer",
            "title": "Request Count"
          },
          "last_request_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Request At"
          },
          "job_to_be_done": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job To Be Done"
          },
          "origin": {
            "$ref": "#/components/schemas/NeedOrigin"
          },
          "requests": {
            "items": {
              "$ref": "#/components/schemas/PortalRequest"
            },
            "type": "array",
            "title": "Requests",
            "description": "Titles only: descriptions can hold other customers' details"
          },
          "supporters": {
            "items": {
              "$ref": "#/components/schemas/PortalSupporter"
            },
            "type": "array",
            "title": "Supporters",
            "description": "Without their reasons, which are theirs to share"
          },
          "updates": {
            "items": {
              "$ref": "#/components/schemas/PortalUpdate"
            },
            "type": "array",
            "title": "Updates",
            "description": "Approved updates written to the requester_id asked for"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "problem",
          "persona",
          "product_area",
          "status",
          "support_count",
          "account_count",
          "request_count",
          "last_request_at",
          "job_to_be_done",
          "origin",
          "requests",
          "supporters",
          "updates"
        ],
        "title": "PortalNeedDetail"
      },
      "PortalNeedPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PortalNeedSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PortalNeedPage"
      },
...
      "PriorityBreakdown": {
        "properties": {
          "priority": {
            "type": "number",
            "title": "Priority",
            "description": "0-100, the sum of contributions"
          },
          "contributions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Contributions",
            "description": "Points per component: 100 x weight x value"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weights",
            "description": "Weights used; renormalised while strategic is unrated"
          },
          "demand": {
            "$ref": "#/components/schemas/DemandOut"
          },
          "urgency": {
            "$ref": "#/components/schemas/UrgencyOut"
          },
          "strategic": {
            "$ref": "#/components/schemas/StrategicOut"
          },
          "quadrant": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "clear_win",
                  "strategic_bet",
                  "popular_off_strategy",
                  "park"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Quadrant",
            "description": "Popular is demand, strategic is S; null while unrated"
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "description": "The PM team that owns the product area"
          }
        },
        "type": "object",
        "required": [
          "priority",
          "contributions",
          "weights",
          "demand",
          "urgency",
          "strategic",
          "quadrant",
          "owner"
        ],
        "title": "PriorityBreakdown"
      },
...
      "RequestCreate": {
        "properties": {
          "requester_id": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 1.0,
            "title": "Requester Id"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9.223372036854776e+18,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id",
            "description": "The customer this is for. Required when staff submit for support, sales or cs."
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "title": "Description",
            "default": ""
          },
          "source": {
            "$ref": "#/components/schemas/RequestSource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requester_id",
          "title",
          "source"
        ],
        "title": "RequestCreate"
      },
      "RequestOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "requester_id": {
            "type": "integer",
            "title": "Requester Id"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "source": {
            "$ref": "#/components/schemas/RequestSource"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "need_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Need Id"
          },
          "problem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "job_to_be_done": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job To Be Done"
          },
          "proposed_solution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Solution"
          },
          "product_area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Area"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "requester_id",
          "account_id",
          "source",
          "title",
          "description",
          "status",
          "need_id",
          "problem",
          "persona",
          "job_to_be_done",
          "proposed_solution",
          "product_area",
          "created_at"
        ],
        "title": "RequestOut"
      },
      "RequestSource": {
        "type": "string",
        "enum": [
          "portal",
          "support",
          "sales",
          "cs",
          "internal"
        ],
        "title": "RequestSource"
      },
      "RequestStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "processed",
          "needs_review"
        ],
        "title": "RequestStatus"
      },
...
      "Segment": {
        "type": "string",
        "enum": [
          "enterprise",
          "mid_market",
          "smb"
        ],
        "title": "Segment"
      },
...
    }
  }
}